competitor: error message when no single results at import
When searching for an existing competitor, it might be that there are more
then one in the DB with the same name and age. This commit does not try to
find the best matching but only provides an error message. The fix (entering
the unique DRV-ID) has to be done manually.

See #14