Add option to move teams between sections
If the team is the last one and is moved down, then it is added on lane 1
of the next race section. If the team is the first one in its section and
is moved up, then it is added on the last lane of the upper race section.

See #33