Sort by time not lane on race details page
When a race section was not started, yet, then the sorting by lane is
automatically done, since the buttons have all same priority and the data
is loaded in order of the lanes.
When a race section is finished, then each line has the priority of the
finisishin time and the table shows automatically the fastest team on top
position.

See #40