show remaning registration time on top of progress bar
When the registration phase starts, the progress bar is very small since
only some minor percentages of the total time frame are over at that time.
As a result, the text was not visible.

With this patch the text is shown on top of the progress bar. This ensures
that it can be fully read by visitors independently of the width of the
progress bar.