add option to undo a cancellation of a team
Even if it sounds easy, UNDO is tricky, because the former position of a team
might be taken by the time when someone clicks "undo". Or the section might
have been started already, so that no-one can be added anymore.

This commits adds a modal popup similar to de-registration and the logic for
finding the first possible slot for a team. For that it looks first in the
original section if there is any free lane and then inspects all other in
ascending order.