automatically de-register in consecutive race as well if possible
Check if any other race points to this consecutive race and if so, check
further if the given team also starts there. In cases where the team does
not start in other races that point to the consecutive race, it can be
de-registered.

NOTE: The UNDO option is not yet implemented.

See #53