add ORDER BY when searching for all races in event
In the "move team to another race" popup all races are shown. Before this
change the races where ordered by database id, what confused the users.

I add the `ORDER BY` to the general requesting method because it does not
produce a lot of overhead on the database site but prevents the fact of
doing the same mistake somewhere else again.