change label logic on `startAll` page
Before the labels where green as soon as there was one checked-in team.
With this commit the labels are red as long as not all relevant teams are
checked in or explicitly marked as "not at start".

Furthermore, the labels show a negative count of how many teams are missing.