add option to start multiple 'ergometer' sections in a single VRA race
This change re-organizes the `startAll` page and divides it into two parts if
there are ergometer races.

As with normal race types, multiple sections can be selected to be started
together. The export file then contains a single race with the lanes/ machines
filled by the given section information.

fixes #35