tokens are added to billing and new overview page for all checked-in tokens
This commit adds a new line to a bill if there are still checked-in tokens in this event for the club.
In addition it adds a new page (and menu entry) for R_REGISTRATION to show all checked-in tokens ordered by club.
This allows to answer questions (who of my members did not return it, yet?) if people do not want to pay.