only show registration manipulation buttons if not a run (in Row&Run) race
A Row&Run race consists of multiple indoor rowing competitions followed by a single run.
Since all competitors start in a mass-start without lane assignment, it is not useful to have the ability to change lane for such races.
In addition, it seems also not useful to have the option for moving a competitor to another race from there.
As a result, these buttons are hidden.

This also leads to a performance boost, because each of the "move" buttons had a DIV in background holding all potential races as list.
After reaching more than 80 registrations, the PHP memory limit of 128MB was exceeded due to that.
By not having the data anymore, the system runs performant again.