tool to combine Concept2 VRA race result files
The Concept2 software splits results into multiple files in different folders.
Each file has another data format and none of them uses a standard.

This tool consists of a folder observer and a processing library.
The first checks all files in the result folders and calculates if a processing is needed for a (new) file combination.
If so, it triggers the main library function which then as a result produces a new file in a given output folder.

See #37