add new page to find the fastest competitors for the whole event
This page shows the top 15 fastest (single) competitors in the event.
The results are separated by gender and race type.
Currently there exists no algorithm to find the one competitor (per gender) who is highest in all categories.

NOTE: The algorithm to find the top competitors currently takes a lot of memory when there are many teams with one person each.
It might be better to combine the two loops and only keep the top 15 times at all time.
However, this will lead to a higher calculation time.

See #60